refactor: Extract shared go2rtc client code into a src/go2rtc module
