+++ src/components-lib/live/providers/go2rtc-experimental/types.ts
@@ -1,29 +1,13 @@
-import { z } from 'zod';
-
+import type {
+  BinaryCallback,
+  Go2RTCMessage,
+  MessageCallback,
+} from '../../../../go2rtc/messages';
 import type {
   MediaLoadedCapabilities,
   MediaTechnology,
   UnsubscribeCallback,
 } from '../../../../types';
-
-// ===========================================================================
-// Control messages
-// ===========================================================================
-
-// go2rtc control messages are JSON text frames of this shape; media flows as
-// separate binary frames.
-export const go2RTCMessageSchema = z.object({
-  type: z.string(),
-
-  // Per-type payload (a codec list, an SDP, an ICE candidate, error text, ...):
-  // absent for some types (e.g. mjpeg) and not always a string, so it is typed
-  // `unknown` and each handler narrows it before use.
-  value: z.unknown().optional(),
-});
-export type Go2RTCMessage = z.infer<typeof go2RTCMessageSchema>;
-
-export type MessageCallback = (message: Go2RTCMessage) => void;
-export type BinaryCallback = (data: ArrayBuffer) => void;
 
 // ===========================================================================
 // Signaling channel
