+++ src/go2rtc/audio.ts
@@ -1,9 +1,9 @@
-import type { EnabledProxyConfig } from '../../../config/schema/common/proxy';
-import { homeAssistantSignAndFetch } from '../../../ha/fetch';
-import type { HomeAssistant } from '../../../ha/types';
-import { createProxiedEndpointIfNecessary } from '../../../ha/web-proxy';
-import type { Endpoint } from '../../../types';
-import { errorToConsole } from '../../../utils/basic';
+import type { EnabledProxyConfig } from '../config/schema/common/proxy';
+import { homeAssistantSignAndFetch } from '../ha/fetch';
+import type { HomeAssistant } from '../ha/types';
+import { createProxiedEndpointIfNecessary } from '../ha/web-proxy';
+import type { Endpoint } from '../types';
+import { errorToConsole } from '../utils/basic';
 import { go2RTCStreamInfoSchema, type Go2RTCStreamInfo } from './types';
 
 const getGo2RTCStreamMetadata = async (
