+++ tests/components-lib/live/providers/go2rtc-experimental/test-utils.ts
@@ -4,12 +4,12 @@ import type {
   MediaSourceFactory,
   MediaSourceInterface,
 } from '../../../../../src/components-lib/live/providers/go2rtc-experimental/adapters/media-source';
+import type { StreamSourceChannel } from '../../../../../src/components-lib/live/providers/go2rtc-experimental/types';
 import type {
   BinaryCallback,
   Go2RTCMessage,
   MessageCallback,
-  StreamSourceChannel,
-} from '../../../../../src/components-lib/live/providers/go2rtc-experimental/types';
+} from '../../../../../src/go2rtc/messages';
 import type { UnsubscribeCallback } from '../../../../../src/types';
 
 // ===========================================================================
@@ -27,32 +27,6 @@ export const SAFARI_17_USER_AGENT =
 // ===========================================================================
 // Fakes for browser APIs jsdom does not provide.
 // ===========================================================================
-
-export class FakeWebSocket extends EventTarget {
-  public binaryType = '';
-  public sent: string[] = [];
-
-  public close = vi.fn();
-  public send = vi.fn((data: string): void => {
-    this.sent.push(data);
-  });
-
-  public asWebSocket(): WebSocket {
-    return this as unknown as WebSocket;
-  }
-
-  public fireOpen(): void {
-    this.dispatchEvent(new Event('open'));
-  }
-
-  public fireClose(): void {
-    this.dispatchEvent(new Event('close'));
-  }
-
-  public fireMessage(data: unknown): void {
-    this.dispatchEvent(new MessageEvent('message', { data }));
-  }
-}
 
 export const createTimeRanges = (ranges: [number, number][]): TimeRanges => ({
   length: ranges.length,
