+++ tests/go2rtc/audio.test.ts
@@ -1,14 +1,14 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { mock } from 'vitest-mock-extended';
 
-import { supports2WayAudio } from '../../../src/camera-manager/utils/go2rtc/audio';
-import { homeAssistantSignAndFetch } from '../../../src/ha/fetch';
-import type { HomeAssistant } from '../../../src/ha/types';
-import { createProxiedEndpointIfNecessary } from '../../../src/ha/web-proxy';
-import type { Endpoint } from '../../../src/types';
+import { supports2WayAudio } from '../../src/go2rtc/audio';
+import { homeAssistantSignAndFetch } from '../../src/ha/fetch';
+import type { HomeAssistant } from '../../src/ha/types';
+import { createProxiedEndpointIfNecessary } from '../../src/ha/web-proxy';
+import type { Endpoint } from '../../src/types';
 
-vi.mock('../../../src/ha/fetch');
-vi.mock('../../../src/ha/web-proxy');
+vi.mock('../../src/ha/fetch');
+vi.mock('../../src/ha/web-proxy');
 
 describe('supports2WayAudio', () => {
   const hass = mock<HomeAssistant>();
